refactor(stringCalculator): switch문 제거, nullAndEmptyValue메서드 이동
