+++ src/main/java/study/Calculator.java
@@ -4,38 +4,25 @@ import java.util.regex.Pattern;
 
 public class Calculator {
 
-
     public int computation(int num1, int num2, String operator) {
-        switch (operator) {
-            case "+":
-                return num1 + num2;
-            case "-":
-                return num1 - num2;
 
-            case "*":
-                return num1 * num2;
-
-            case "/":
-                return num1 / num2;
+        if (operator.equals("+")) {
+            return num1 + num2;
+        } else if (operator.equals("-")) {
+            return num1 - num2;
+        } else if (operator.equals("*")) {
+            return num1 * num2;
+        } else if (operator.equals("/")) {
+            return num1 / num2;
         }
 
         return 0;
     }
 
-    public void nullAndEmptyValue(String number) {
-        if (number == null || number.equals("")) {
-            throw new IllegalArgumentException();
-        }
-    }
-
     public String checkPolicyOperator(String operator) {
-
-        switch (operator) {
-            case "+":
-            case "-":
-            case "*":
-            case "/":
-                return operator;
+        boolean isContainsOperator = "+_*/".contains(operator);
+        if (isContainsOperator) {
+            return operator;
         }
 
         throw new IllegalArgumentException();
@@ -45,7 +32,7 @@ public class Calculator {
         return formula.split(" ");
     }
 
-    public boolean isNumber(String value){
+    public boolean isNumber(String value) {
         return Pattern.matches("^[0-9]*$", value);
     }
 }
